+++ changelog.md
@@ -1,3 +1,8 @@
 # Changes
-- Add capability check for Ender Skills to remove XP from the Ender Skills book. Can't let people squirrel away any xp anywhere!
-- Also reset Ender Skills cooldown and endurance.
+- Added config options for rerolling on death, starting locked, cooldown timer for using the dice, sending your inventory into a chest, whether ender chest resets, etc.
+- Added new Dice item that can be used to reroll.
+- Modified initial inventory to include the dice.
+- Set new defaults for config options.
+
+# Fixes
+- Fixed issue where useCurrentDim and overrideDim weren't working. They should be now (unless I broke something since then).